requirements.txt
openai
google-generativeai
groq
# torch
# diffusers
# transformers
# accelerate
# streamlit
